sqrt

+++ recursion/5-sqrt_recursion.c
@@ -6,13 +6,13 @@
  */
 int _sqrt_recursion(int n)
 {
-	long int i = 1;
+	int i = 1;
 
 	if (i * i <= n)
 	{
 		if (i * i == n)
 			return (i);
-	_sqrt_recursion(i + 1);
+		_sqrt_recursion(i + 1);
 	}
 	return (-1);
 }
